pyproject.toml
[project]
name = "godot-sandbox-tools"
version = "0.1.0"
description = "Utility scripts for inspecting and preparing Penguin Dash assets."
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2,<12",
]

[tool.uv]
package = false
